ai/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "AI сервис для автоматической разметки PDF научных статей"
authors = [
    {name = "Dobrosvet",email = "dimaprokofev@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "transformers>=4.35.0",
    "torch>=2.1.0",
    "accelerate>=0.24.0",
    "sentencepiece>=0.1.99",
    "protobuf>=4.25.0",
    "PyPDF2>=3.0.1",
    "pdfplumber>=0.10.0",
    "Pillow>=10.1.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "aiofiles>=23.2.0",
    "python-multipart>=0.0.6",
    "grpcio>=1.59.0",
    "grpcio-tools>=1.59.0",
    "neomodel>=5.3.0",
    "aioboto3>=12.0.0",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.1.0",
    "loguru>=0.7.2",
    "tiktoken>=0.5.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.2"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "ai"
version = "0.1.0"
description = "AI сервис для автоматической разметки PDF научных статей"
authors = ["Dobrosvet <dimaprokofev@yandex.ru>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
transformers = "^4.35.0"
torch = "^2.1.0"
accelerate = "^0.24.0"
sentencepiece = "^0.1.99"
protobuf = "^4.25.0"
PyPDF2 = "^3.0.1"
pdfplumber = "^0.10.0"
Pillow = "^10.1.0"
opencv-python = "^4.8.0"
numpy = "^1.24.0"
requests = "^2.31.0"
aiofiles = "^23.2.0"
python-multipart = "^0.0.6"
grpcio = "^1.59.0"
grpcio-tools = "^1.59.0"
neomodel = "^5.3.0"
aioboto3 = "^12.0.0"
boto3 = "^1.34.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"
loguru = "^0.7.2"
tiktoken = "^0.5.0"
langchain = "^0.1.0"
langchain-community = "^0.0.10"
chromadb = "^0.4.0"
sentence-transformers = "^2.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.7.0"
